undo test locator class extraction
